requirements.txt
flask
openai
line-bot-sdk
python-dotenv
